104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,18 +10,18 @@
 int main(void)
 {
     int num;
-    unsigned long int fib1 = 1, fib2 = 2, next;
+    unsigned long long int fib1 = 1, fib2 = 2, next;
 
-    printf("%lu", fib1);
+    printf("%llu", fib1);
 
     for (num = 1; num < 98; num++)
     {
         if (num == 1)
-            printf(", %lu", fib2);
+            printf(", %llu", fib2);
         else
         {
             next = fib1 + fib2;
-            printf(", %lu", next);
+            printf(", %llu", next);
 
             fib1 = fib2;
             fib2 = next;
